+++ tests/advanced_control_flow/main.rs
@@ -395,7 +395,7 @@ fn main() {
         print!("Found center at [{}][{}]: {}\n", rowIdx, colIdx, cell);
         continue
     }
-        if { let __tmp_x = rowIdx; let __tmp_y = 2; __tmp_x == __tmp_y } && { let __tmp_x = colIdx; let __tmp_y = 2; __tmp_x == __tmp_y } {
+        if { let __tmp_x = rowIdx as i32; let __tmp_y = 2; __tmp_x == __tmp_y } && { let __tmp_x = colIdx as i32; let __tmp_y = 2; __tmp_x == __tmp_y } {
         print!("Last cell [{}][{}]: {}\n", rowIdx, colIdx, cell);
         break
     }
@@ -493,7 +493,7 @@ fn main() {
     let mut testData = Arc::new(Mutex::new(Some(vec![vec![1, 2, 3], vec![], vec![1, -2, 3], vec![1, 200, 3], vec![10, 20, 30]])));
 
     { let __range_holder = testData.clone(); let __range_guard = __range_holder.lock().unwrap(); let __range_values = __range_guard.as_ref().map(|__v| __v.as_slice()).unwrap_or(&[]); for (i, data) in __range_values.iter().enumerate() {
-        print!("Testing dataset {}: {}\n", { let __tmp_x = i; let __tmp_y = 1; __tmp_x + __tmp_y }, format_slice_values(data));
+        print!("Testing dataset {}: {}\n", { let __tmp_x = i as i32; let __tmp_y = 1; __tmp_x + __tmp_y }, format_slice_values(data));
         let mut err = { let __f_ptr: *mut Box<dyn FnMut(Arc<Mutex<Option<Vec<i32>>>>) -> Arc<Mutex<Option<Box<dyn StdError + Send + Sync>>>> + Send + Sync> = { let mut __f_guard = processData.lock().unwrap(); __f_guard.as_mut().unwrap() as *mut Box<dyn FnMut(Arc<Mutex<Option<Vec<i32>>>>) -> Arc<Mutex<Option<Box<dyn StdError + Send + Sync>>>> + Send + Sync> }; let __f = unsafe { &mut *__f_ptr }; (*__f)(Arc::new(Mutex::new(Some(data.clone())))) };
     if (*err.lock().unwrap()).is_some() {
         print!("  Error: {}\n", format!("{}", (*err.lock().unwrap().as_ref().unwrap())));
